beta process
