+++ src/javascript/binary/websocket_pages/trade/beta/duration.js
@@ -1,5 +1,6 @@
 var Barriers_Beta    = require('./barriers').Barriers_Beta;
 var Contract_Beta    = require('./contract').Contract_Beta;
+var Price_Beta       = require('./price').Price_Beta;
 var Defaults         = require('../defaults').Defaults;
 var moment           = require('moment');
 var Content          = require('../../../common_functions/content').Content;
@@ -8,6 +9,7 @@ var isVisible        = require('../../../common_functions/common_functions').isV
 var durationOrder    = require('../common').durationOrder;
 var selectOption     = require('../common').selectOption;
 var timeIsValid      = require('../common').timeIsValid;
+var showPriceOverlay = require('../common').showPriceOverlay;
 var DatePicker       = require('../../../components/date_picker').DatePicker;
 var toReadableFormat = require('../../../common_functions/string_util').toReadableFormat;
 var toISOFormat      = require('../../../common_functions/string_util').toISOFormat;
@@ -25,9 +27,7 @@ var toISOFormat      = require('../../../common_functions/string_util').toISOFor
 var Durations_Beta = (function() {
     'use strict';
 
-    var trading_times     = {},
-        selected_duration = {},
-        expiry_time_def   = '',
+    var selected_duration = {},
         has_end_date      = 0;
 
     var displayDurations = function() {
@@ -317,28 +317,6 @@ var Durations_Beta = (function() {
         target.appendChild(fragment);
     };
 
-    var processTradingTimesAnswer = function(response) {
-        if (!trading_times.hasOwnProperty(response.echo_req.trading_times) && response.hasOwnProperty('trading_times') && response.trading_times.hasOwnProperty('markets')) {
-            for (var i = 0; i < response.trading_times.markets.length; i++) {
-                var submarkets = response.trading_times.markets[i].submarkets;
-                if (submarkets) {
-                    for (var j = 0; j < submarkets.length; j++) {
-                        var symbols = submarkets[j].symbols;
-                        if (symbols) {
-                            for (var k = 0; k < symbols.length; k++) {
-                                var symbol = symbols[k];
-                                if (!trading_times[response.echo_req.trading_times]) {
-                                    trading_times[response.echo_req.trading_times] = {};
-                                }
-                                trading_times[response.echo_req.trading_times][symbol.symbol] = symbol.times.close;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    };
-
     var selectEndDate = function(end_date) {
         var expiry_time = document.getElementById('expiry_time'),
             date_start  = document.getElementById('date_start'),
@@ -364,7 +342,7 @@ var Durations_Beta = (function() {
             Durations_Beta.setTime(expiry_time.value);
             Defaults.set('expiry_time', Defaults.get('expiry_time') || expiry_time.value);
             expiry_time.show();
-            processPriceRequest_Beta();
+            Price_Beta.processPriceRequest_Beta();
         }
 
         Barriers_Beta.display();
@@ -419,18 +397,27 @@ var Durations_Beta = (function() {
         }
     };
 
+    function processTradingTimesRequest_Beta(date) {
+        var trading_times = Durations_Beta.trading_times();
+        if (trading_times.hasOwnProperty(date)) {
+            Price_Beta.processPriceRequest_Beta();
+        } else {
+            showPriceOverlay();
+            BinarySocket.send({
+                trading_times: date,
+            });
+        }
+    }
+
     return {
         display                  : displayDurations,
         displayEndTime           : displayEndTime,
         populate                 : durationPopulate,
-        processTradingTimesAnswer: processTradingTimesAnswer,
         selectEndDate            : selectEndDate,
         validateMinDurationAmount: validateMinDurationAmount,
         onStartDateChange        : onStartDateChange,
 
-        setTime      : function(time) { $('#expiry_time').val(time); Defaults.set('expiry_time', time); expiry_time_def = time; },
-        getTime      : function() { return expiry_time_def; },
-        trading_times: function() { return trading_times; },
+        setTime      : function(time) { $('#expiry_time').val(time); Defaults.set('expiry_time', time); },
         select_amount: function(a) { selected_duration.amount = a; },
         select_unit  : function(u) { selected_duration.unit = u; },
     };
