+++ src/javascript/binary/websocket_pages/trade/beta/event.js
@@ -3,7 +3,11 @@ var Barriers_Beta                  = require('./barriers').Barriers_Beta;
 var Contract_Beta                  = require('./contract').Contract_Beta;
 var Durations_Beta                 = require('./duration').Durations_Beta;
 var Price_Beta                     = require('./price').Price_Beta;
-var StartDates_Beta                = require('./starttime').StartDates_Beta;
+var processMarket_Beta             = require('./process').processMarket_Beta;
+var processContractForm_Beta       = require('./process').processContractForm_Beta;
+var processForgetTicks_Beta        = require('./process').processForgetTicks_Beta;
+var onExpiryTypeChange             = require('./process').onExpiryTypeChange;
+var onDurationUnitChange           = require('./process').onDurationUnitChange;
 var Defaults                       = require('../defaults').Defaults;
 var Tick                           = require('../tick').Tick;
 var onlyNumericOnKeypress          = require('../../../common_functions/event_handler').onlyNumericOnKeypress;
@@ -20,6 +24,7 @@ var reloadPage                     = require('../common').reloadPage;
 var chartFrameSource               = require('../common').chartFrameSource;
 var displayTooltip_Beta            = require('../common').displayTooltip_Beta;
 var timeIsValid                    = require('../common').timeIsValid;
+var getStartDateNode               = require('../common_independent').getStartDateNode;
 var isVisible                      = require('../../../common_functions/common_functions').isVisible;
 var dateValueChanged               = require('../../../common_functions/common_functions').dateValueChanged;
 var TimePicker                     = require('../../../components/time_picker').TimePicker;
@@ -34,52 +39,6 @@ var TimePicker                     = require('../../../components/time_picker').
  */
 var TradingEvents_Beta = (function () {
     'use strict';
-
-    var onExpiryTypeChange = function(value) {
-        if (!value || !$('#expiry_type').find('option[value=' + value + ']').length) {
-            value = 'duration';
-        }
-        $('#expiry_type').val(value);
-
-        var make_price_request = 0;
-        if (value === 'endtime') {
-            Durations_Beta.displayEndTime();
-            if (Defaults.get('expiry_date')) {
-                Durations_Beta.selectEndDate(moment(Defaults.get('expiry_date')));
-                make_price_request = -1;
-            }
-            Defaults.remove('duration_units', 'duration_amount');
-        } else {
-            StartDates_Beta.enable();
-            Durations_Beta.display();
-            if (Defaults.get('duration_units')) {
-                TradingEvents_Beta.onDurationUnitChange(Defaults.get('duration_units'));
-            }
-            var duration_amount = Defaults.get('duration_amount');
-            if (duration_amount && duration_amount > $('#duration_minimum').text()) {
-                $('#duration_amount').val(duration_amount);
-            }
-            make_price_request = 1;
-            Defaults.remove('expiry_date', 'expiry_time', 'end_date');
-            Durations_Beta.validateMinDurationAmount();
-        }
-
-        return make_price_request;
-    };
-
-    var onDurationUnitChange = function(value) {
-        if (!value || !$('#duration_units').find('option[value=' + value + ']').length) {
-            return 0;
-        }
-
-        $('#duration_units').val(value);
-        Defaults.set('duration_units', value);
-
-        Durations_Beta.select_unit(value);
-        Durations_Beta.populate();
-
-        return 1;
-    };
 
     var initiate = function () {
         /*
@@ -180,7 +139,7 @@ var TradingEvents_Beta = (function () {
             }
             Defaults.set('duration_amount', e.target.value);
             Durations_Beta.select_amount(e.target.value);
-            processPriceRequest_Beta();
+            Price_Beta.processPriceRequest_Beta();
             submitForm(document.getElementById('websocket_form'));
         }
         var durationAmountElement = document.getElementById('duration_amount'),
@@ -213,7 +172,7 @@ var TradingEvents_Beta = (function () {
             expiryTypeElement.addEventListener('change', function(e) {
                 Defaults.set('expiry_type', e.target.value);
                 onExpiryTypeChange(e.target.value);
-                if (expiryTypeElement.value !== 'endtime') processPriceRequest_Beta();
+                if (expiryTypeElement.value !== 'endtime') Price_Beta.processPriceRequest_Beta();
             });
         }
 
@@ -224,8 +183,8 @@ var TradingEvents_Beta = (function () {
         if (durationUnitElement) {
             durationUnitElement.addEventListener('change', function (e) {
                 Defaults.remove('barrier', 'barrier_high', 'barrier_low');
-                TradingEvents_Beta.onDurationUnitChange(e.target.value);
-                processPriceRequest_Beta();
+                onDurationUnitChange(e.target.value);
+                Price_Beta.processPriceRequest_Beta();
             });
         }
 
@@ -259,7 +218,7 @@ var TradingEvents_Beta = (function () {
                     // if start time is less than end time
                     if (timeIsValid($('#expiry_time'))) {
                         Durations_Beta.setTime(endTimeElement.value);
-                        processPriceRequest_Beta();
+                        Price_Beta.processPriceRequest_Beta();
                     }
                     return true;
                 });
@@ -278,7 +237,7 @@ var TradingEvents_Beta = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('amount', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -288,13 +247,13 @@ var TradingEvents_Beta = (function () {
          * whether start time is forward starting or not and request
          * new price
          */
-        var dateStartElement = StartDates_Beta.node();
+        var dateStartElement = getStartDateNode();
         if (dateStartElement) {
             dateStartElement.addEventListener('change', function (e) {
                 Defaults.set('date_start', e.target.value);
                 var r = Durations_Beta.onStartDateChange(e.target.value);
                 if (r >= 0) {
-                    processPriceRequest_Beta();
+                    Price_Beta.processPriceRequest_Beta();
                 }
             });
         }
@@ -307,7 +266,7 @@ var TradingEvents_Beta = (function () {
         if (amountTypeElement) {
             amountTypeElement.addEventListener('change', function (e) {
                 Defaults.set('amount_type', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
             });
         }
 
@@ -350,7 +309,7 @@ var TradingEvents_Beta = (function () {
                 if (stopTypeDollarLabel && isVisible(stopTypeDollarLabel)) {
                     stopTypeDollarLabel.textContent = e.target.value;
                 }
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
             });
         }
 
@@ -376,7 +335,7 @@ var TradingEvents_Beta = (function () {
                 if (id && askPrice) {
                     BinarySocket.send(params);
                     Price_Beta.incrFormId();
-                    processForgetProposals_Beta();
+                    Price_Beta.processForgetProposals_Beta();
                 }
             }
         });
@@ -389,7 +348,7 @@ var TradingEvents_Beta = (function () {
                 e.preventDefault();
                 document.getElementById('contract_confirmation_container').style.display = 'none';
                 document.getElementById('contracts_list').style.display = 'flex';
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
             }
         });
 
@@ -401,7 +360,7 @@ var TradingEvents_Beta = (function () {
             barrierElement.addEventListener('input', debounce(function (e) {
                 Barriers_Beta.validateBarrier();
                 Defaults.set('barrier', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }, 1000));
         }
@@ -413,7 +372,7 @@ var TradingEvents_Beta = (function () {
         if (lowBarrierElement) {
             lowBarrierElement.addEventListener('input', debounce(function (e) {
                 Defaults.set('barrier_low', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -425,7 +384,7 @@ var TradingEvents_Beta = (function () {
         if (highBarrierElement) {
             highBarrierElement.addEventListener('input', debounce(function (e) {
                 Defaults.set('barrier_high', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -437,7 +396,7 @@ var TradingEvents_Beta = (function () {
         if (predictionElement) {
             predictionElement.addEventListener('change', debounce(function (e) {
                 Defaults.set('prediction', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -452,7 +411,7 @@ var TradingEvents_Beta = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('amount_per_point', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -462,7 +421,7 @@ var TradingEvents_Beta = (function () {
          */
         var stopTypeEvent = function (e) {
             Defaults.set('stop_type', e.target.value);
-            processPriceRequest_Beta();
+            Price_Beta.processPriceRequest_Beta();
         };
 
         var stopTypeElement = document.querySelectorAll('input[name="stop_type"]');
@@ -482,7 +441,7 @@ var TradingEvents_Beta = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('stop_loss', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -497,7 +456,7 @@ var TradingEvents_Beta = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('stop_profit', e.target.value);
-                processPriceRequest_Beta();
+                Price_Beta.processPriceRequest_Beta();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -540,9 +499,7 @@ var TradingEvents_Beta = (function () {
     }
 
     return {
-        init                : initiate,
-        onExpiryTypeChange  : onExpiryTypeChange,
-        onDurationUnitChange: onDurationUnitChange,
+        init: initiate,
     };
 })();
