+++ src/javascript/binary/websocket_pages/trade/beta/message.js
@@ -1,5 +1,11 @@
 var TradingAnalysis_Beta = require('./analysis').TradingAnalysis_Beta;
 var Purchase_Beta        = require('./purchase').Purchase_Beta;
+var processActiveSymbols_Beta = require('./process').processActiveSymbols_Beta;
+var processContract_Beta      = require('./process').processContract_Beta;
+var forgetTradingStreams_Beta = require('./process').forgetTradingStreams_Beta;
+var processTick_Beta          = require('./process').processTick_Beta;
+var processProposal_Beta      = require('./process').processProposal_Beta;
+var processTradingTimes_Beta  = require('./process').processTradingTimes_Beta;
 var displayCurrencies = require('../currency').displayCurrencies;
 var Notifications     = require('../notifications').Notifications;
 var Symbols           = require('../symbols').Symbols;
