+++ src/javascript/binary/websocket_pages/trade/beta/process.js
@@ -2,7 +2,6 @@ var TradingAnalysis_Beta           = require('./analysis').TradingAnalysis_Beta;
 var Barriers_Beta                  = require('./barriers').Barriers_Beta;
 var Contract_Beta                  = require('./contract').Contract_Beta;
 var Durations_Beta                 = require('./duration').Durations_Beta;
-var TradingEvents_Beta             = require('./event').TradingEvents_Beta;
 var Price_Beta                     = require('./price').Price_Beta;
 var Purchase_Beta                  = require('./purchase').Purchase_Beta;
 var StartDates_Beta                = require('./starttime').StartDates_Beta;
@@ -13,7 +12,6 @@ var Tick                           = require('../tick').Tick;
 var State                          = require('../../../base/storage').State;
 var displayUnderlyings             = require('../common').displayUnderlyings;
 var setFormPlaceholderContent_Beta = require('../set_values').setFormPlaceholderContent_Beta;
-var showPriceOverlay               = require('../common').showPriceOverlay;
 var hidePriceOverlay               = require('../common').hidePriceOverlay;
 var hideFormOverlay                = require('../common').hideFormOverlay;
 var showFormOverlay                = require('../common').showFormOverlay;
@@ -25,6 +23,8 @@ var updateWarmChart                = require('../common').updateWarmChart;
 var displayContractForms           = require('../common').displayContractForms;
 var displayMarkets                 = require('../common').displayMarkets;
 var displayTooltip_Beta            = require('../common').displayTooltip_Beta;
+var processTradingTimesAnswer      = require('../common_independent').processTradingTimesAnswer;
+var moment = require('moment');
 
 /*
  * This function process the active symbols to get markets
@@ -120,7 +120,7 @@ function processContract_Beta(contracts) {
     'use strict';
 
     if (contracts.hasOwnProperty('error') && contracts.error.code === 'InvalidSymbol') {
-        processForgetProposals_Beta();
+        Price_Beta.processForgetProposals_Beta();
         var container = document.getElementById('contract_confirmation_container'),
             message_container = document.getElementById('confirmation_message'),
             confirmation_error = document.getElementById('confirmation_error'),
@@ -204,10 +204,10 @@ function processContractForm_Beta() {
     if (Defaults.get('currency')) selectOption(Defaults.get('currency'), document.getElementById('currency'));
 
     var expiry_type = Defaults.get('expiry_type') || 'duration';
-    var make_price_request = TradingEvents_Beta.onExpiryTypeChange(expiry_type);
+    var make_price_request = onExpiryTypeChange(expiry_type);
 
     if (make_price_request >= 0) {
-        processPriceRequest_Beta();
+        Price_Beta.processPriceRequest_Beta();
     }
 
     if (Defaults.get('formname') === 'spreads') {
@@ -276,58 +276,8 @@ function displaySpreads_Beta() {
 }
 
 function forgetTradingStreams_Beta() {
-    processForgetProposals_Beta();
+    Price_Beta.processForgetProposals_Beta();
     processForgetTicks_Beta();
-}
-/*
- * Function to request for cancelling the current price proposal
- */
-function processForgetProposals_Beta() {
-    'use strict';
-
-    showPriceOverlay();
-    BinarySocket.send({
-        forget_all: 'proposal',
-    });
-    Price_Beta.clearMapping();
-}
-
-/*
- * Function to process and calculate price based on current form
- * parameters or change in form parameters
- */
-function processPriceRequest_Beta() {
-    'use strict';
-
-    Price_Beta.incrFormId();
-    processForgetProposals_Beta();
-    showPriceOverlay();
-    var types = Contract_Beta.contractType()[Contract_Beta.form()];
-    if (Contract_Beta.form() === 'digits') {
-        switch (sessionStorage.getItem('formname')) {
-            case 'matchdiff':
-                types = {
-                    DIGITMATCH: 1,
-                    DIGITDIFF : 1,
-                };
-                break;
-            case 'evenodd':
-                types = {
-                    DIGITEVEN: 1,
-                    DIGITODD : 1,
-                };
-                break;
-            case 'overunder':
-                types = {
-                    DIGITOVER : 1,
-                    DIGITUNDER: 1,
-                };
-            // no default
-        }
-    }
-    Object.keys(types).forEach(function(typeOfContract) {
-        BinarySocket.send(Price_Beta.proposal(typeOfContract));
-    });
 }
 
 /*
@@ -380,35 +330,67 @@ function processProposal_Beta(response) {
     }
 }
 
-function processTradingTimesRequest_Beta(date) {
-    var trading_times = Durations_Beta.trading_times();
-    if (trading_times.hasOwnProperty(date)) {
-        processPriceRequest_Beta();
-    } else {
-        showPriceOverlay();
-        BinarySocket.send({
-            trading_times: date,
-        });
-    }
-}
-
 function processTradingTimes_Beta(response) {
-    Durations_Beta.processTradingTimesAnswer(response);
-
-    processPriceRequest_Beta();
+    processTradingTimesAnswer(response);
+    Price_Beta.processPriceRequest_Beta();
 }
+
+var onExpiryTypeChange = function(value) {
+    if (!value || !$('#expiry_type').find('option[value=' + value + ']').length) {
+        value = 'duration';
+    }
+    $('#expiry_type').val(value);
+
+    var make_price_request = 0;
+    if (value === 'endtime') {
+        Durations_Beta.displayEndTime();
+        if (Defaults.get('expiry_date')) {
+            Durations_Beta.selectEndDate(moment(Defaults.get('expiry_date')));
+            make_price_request = -1;
+        }
+        Defaults.remove('duration_units', 'duration_amount');
+    } else {
+        StartDates_Beta.enable();
+        Durations_Beta.display();
+        if (Defaults.get('duration_units')) {
+            onDurationUnitChange(Defaults.get('duration_units'));
+        }
+        var duration_amount = Defaults.get('duration_amount');
+        if (duration_amount && duration_amount > $('#duration_minimum').text()) {
+            $('#duration_amount').val(duration_amount);
+        }
+        make_price_request = 1;
+        Defaults.remove('expiry_date', 'expiry_time', 'end_date');
+        Durations_Beta.validateMinDurationAmount();
+    }
+
+    return make_price_request;
+};
+
+var onDurationUnitChange = function(value) {
+    if (!value || !$('#duration_units').find('option[value=' + value + ']').length) {
+        return 0;
+    }
+
+    $('#duration_units').val(value);
+    Defaults.set('duration_units', value);
+
+    Durations_Beta.select_unit(value);
+    Durations_Beta.populate();
+
+    return 1;
+};
 
 module.exports = {
-    processActiveSymbols_Beta      : processActiveSymbols_Beta,
-    processMarket_Beta             : processMarket_Beta,
-    processContract_Beta           : processContract_Beta,
-    processContractForm_Beta       : processContractForm_Beta,
-    forgetTradingStreams_Beta      : forgetTradingStreams_Beta,
-    processForgetProposals_Beta    : processForgetProposals_Beta,
-    processPriceRequest_Beta       : processPriceRequest_Beta,
-    processForgetTicks_Beta        : processForgetTicks_Beta,
-    processTick_Beta               : processTick_Beta,
-    processProposal_Beta           : processProposal_Beta,
-    processTradingTimesRequest_Beta: processTradingTimesRequest_Beta,
-    processTradingTimes_Beta       : processTradingTimes_Beta,
+    processActiveSymbols_Beta: processActiveSymbols_Beta,
+    processMarket_Beta       : processMarket_Beta,
+    processContract_Beta     : processContract_Beta,
+    processContractForm_Beta : processContractForm_Beta,
+    forgetTradingStreams_Beta: forgetTradingStreams_Beta,
+    processForgetTicks_Beta  : processForgetTicks_Beta,
+    processTick_Beta         : processTick_Beta,
+    processProposal_Beta     : processProposal_Beta,
+    processTradingTimes_Beta : processTradingTimes_Beta,
+    onExpiryTypeChange       : onExpiryTypeChange,
+    onDurationUnitChange     : onDurationUnitChange,
 };
