+++ src/javascript/binary/websocket_pages/trade/duration.js
@@ -10,7 +10,7 @@ var durationOrder    = require('./common').durationOrder;
 var selectOption     = require('./common').selectOption;
 var timeIsValid      = require('./common').timeIsValid;
 var showPriceOverlay = require('./common').showPriceOverlay;
-var getTradingTimes = require('./common_independent').getTradingTimes;
+var getTradingTimes  = require('./common_independent').getTradingTimes;
 var DatePicker       = require('../../components/date_picker').DatePicker;
 var toReadableFormat = require('../../common_functions/string_util').toReadableFormat;
 var toISOFormat      = require('../../common_functions/string_util').toISOFormat;
