+++ src/javascript/index.js
@@ -26,6 +26,3 @@ require('jquery.scrollto');
 require('./binary/components/trackjs_onerror');
 require('./binary/static_pages/static_pjax');
 require('./binary/websocket_pages/websocket_pjax');
-
-// circular dependency
-exportAllFunctions(require('./binary/websocket_pages/trade/beta/process'));
